[bug 681] Plugin Submission + Generic Plugin Utility Functions & KeyBinding Feature: Mouseless / Mousefree / Keyboard URL opening or yanking #681
- removed plugin_util
- added key unbind feature
- updated plugin helper
- updated plugin helper - to have config as part of constructor
- updated MouseFreeURLHandler for plugin.py, unload
- removed circular dependency between plugin.py and config.py due to KeyBindUtil due to merging of plugin_util.py
